requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.6.2
Automat==0.7.0
breadability==0.1.20
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
docopt==0.6.2
Flask==1.0.3
future==0.17.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.4
MarkupSafe==1.1.1
nltk==3.4.3
numpy==1.16.4
oauthlib==3.0.1
parsedatetime==2.4
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycountry==18.12.8
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
service-identity==18.1.0
six==1.12.0
sumy==0.8.1
Twisted==19.2.1
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.3
watson-developer-cloud==1.7.1
websocket-client==0.48.0
Werkzeug==0.15.4
zope.interface==4.6.0
